feat? remove a bunch of unused code and reorganise fns
